[refactor] 重构 interpreter 部分代码
