+++ hybridclr/interpreter/InterpreterModule.cpp
@@ -14,6 +14,7 @@
 #include "../transform/Transform.h"
 
 #include "MethodBridge.h"
+#include "InterpreterUtil.h"
 
 namespace hybridclr
 {
@@ -306,7 +307,7 @@ namespace hybridclr
 		{
 			return (InterpMethodInfo*)methodInfo->interpData;
 		}
-		Interpreter::RuntimeClassCCtorInit(methodInfo);
+		RuntimeClassCCtorInit(methodInfo);
 		IL2CPP_ASSERT(methodInfo->isInterpterImpl);
 
 		metadata::Image* image = metadata::IsInterpreterMethod(methodInfo) ? hybridclr::metadata::MetadataModule::GetImage(methodInfo->klass)
