+++ hybridclr/interpreter/Interpreter_Execute.cpp
@@ -14,12 +14,14 @@
 #include "vm/Runtime.h"
 #include "metadata/GenericMetadata.h"
 
+#include "../metadata/MetadataModule.h"
+
 #include "Instruction.h"
 #include "MethodBridge.h"
 #include "InstrinctDef.h"
 #include "MemoryUtil.h"
-#include "../metadata/MetadataModule.h"
 #include "InterpreterModule.h"
+#include "InterpreterUtil.h"
 
 using namespace hybridclr::metadata;
 
@@ -27,6 +29,7 @@ namespace hybridclr
 {
 namespace interpreter
 {
+
 
 #pragma region memory
 
@@ -852,41 +855,6 @@ namespace interpreter
 		uint32_t size = klass->castClass->instance_size - sizeof(Il2CppObject);
 		return *(GetNulllableHasValueOffset(nullableValueObj, size));
 	}
-
-	/*
-	inline void GetNullableValueOrDefault(void* dst, void* nullableValueObj, Il2CppClass* klass)
-	{
-		uint32_t size = klass->castClass->instance_size - sizeof(Il2CppObject);
-		if (*GetNulllableHasValueOffset(nullableValueObj, size))
-		{
-			std::memmove(dst, nullableValueObj, size);
-		}
-		else
-		{
-			std::memset(dst, 0, size);
-		}
-	}
-
-	inline void GetNullableValueOrDefault(void* dst, void* nullableValueObj, void* defaultData, Il2CppClass* klass)
-	{
-		uint32_t size = klass->castClass->instance_size - sizeof(Il2CppObject);
-		std::memmove(dst, *GetNulllableHasValueOffset(nullableValueObj, size) ? nullableValueObj : defaultData, size);
-	}
-
-	inline void GetNullableValue(void* dst, void* nullableValueObj, Il2CppClass* klass)
-	{
-		uint32_t size = klass->castClass->instance_size - sizeof(Il2CppObject);
-		if (*GetNulllableHasValueOffset(nullableValueObj, size))
-		{
-			std::memmove(dst, nullableValueObj, size);
-		}
-		else
-		{
-			il2cpp::vm::Exception::Raise(il2cpp::vm::Exception::GetInvalidOperationException("Nullable object must have a value."));
-		}
-	}
-	*/
-	// ExpandLocationData2StackDataByType
 	
 	inline void GetNullableValueOrDefault2StackDataByType(void* dst, void* nullableValueObj, Il2CppClass* klass)
 	{
@@ -1224,7 +1192,7 @@ inline void CallDelegateMethod(uint16_t invokeParamCount, const MethodInfo * met
 		}
 		else
 		{
-			Interpreter::RuntimeClassCCtorInit(method);
+			RuntimeClassCCtorInit(method);
 			staticM2NMethod(method, argIdxs + 1, localVarBase, ret);
 		}
 	}
@@ -4694,7 +4662,7 @@ else \
 					uint32_t __methodInfo = *(uint32_t*)(ip + 8);
 					uint32_t __argIdxs = *(uint32_t*)(ip + 12);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__methodInfo]);
-					Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+					RuntimeClassCCtorInit(_resolvedMethod);
 				    ((Managed2NativeCallMethod)imi->resolveDatas[__managed2NativeMethod])(_resolvedMethod, ((uint16_t*)&imi->resolveDatas[__argIdxs]), localVarBase, nullptr);
 				    ip += 16;
 				    continue;
@@ -4706,7 +4674,7 @@ else \
 					uint32_t __argIdxs = *(uint32_t*)(ip + 12);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__methodInfo]);
-					Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+					RuntimeClassCCtorInit(_resolvedMethod);
 				    void* _ret = (void*)(localVarBase + __ret);
 				    ((Managed2NativeCallMethod)imi->resolveDatas[__managed2NativeMethod])(_resolvedMethod, ((uint16_t*)&imi->resolveDatas[__argIdxs]), localVarBase, _ret);
 				    ip += 16;
@@ -4720,7 +4688,7 @@ else \
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 					uint8_t __retLocationType = *(uint8_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__methodInfo]);
-					Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+					RuntimeClassCCtorInit(_resolvedMethod);
 				    void* _ret = (void*)(localVarBase + __ret);
 				    ((Managed2NativeCallMethod)imi->resolveDatas[__managed2NativeMethod])(_resolvedMethod, ((uint16_t*)&imi->resolveDatas[__argIdxs]), localVarBase, _ret);
 				    ExpandLocationData2StackDataByType(_ret, (LocationDataType)__retLocationType);
@@ -4999,7 +4967,7 @@ else \
 										_argBasePtr++;
 										_resolvedArgIdxs++;
 										_managed2NativeCallMethod = _staticM2NMethod;
-										Interpreter::RuntimeClassCCtorInit(method);
+										RuntimeClassCCtorInit(method);
 									}
 									break;
 								}
@@ -5129,7 +5097,7 @@ else \
 										_argBasePtr++;
 										_resolvedArgIdxs++;
 										_managed2NativeCallMethod = _staticM2NMethod;
-										Interpreter::RuntimeClassCCtorInit(method);
+										RuntimeClassCCtorInit(method);
 									}
 									break;
 								}
@@ -5260,7 +5228,7 @@ else \
 										_argBasePtr++;
 										_resolvedArgIdxs++;
 										_managed2NativeCallMethod = _staticM2NMethod;
-										Interpreter::RuntimeClassCCtorInit(method);
+										RuntimeClassCCtorInit(method);
 									}
 									break;
 								}
@@ -6955,7 +6923,7 @@ else \
 				{
 					uint32_t __method = *(uint32_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -6966,7 +6934,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int8_t(*_NativeMethod_)(MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -6977,7 +6945,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -6988,7 +6956,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int16_t(*_NativeMethod_)(MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -6999,7 +6967,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint16_t(*_NativeMethod_)(MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -7010,7 +6978,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -7021,7 +6989,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -7032,7 +7000,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -7043,7 +7011,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -7054,7 +7022,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef HtVector2f(*_NativeMethod_)(MethodInfo*);
 				    *(HtVector2f*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -7065,7 +7033,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef HtVector3f(*_NativeMethod_)(MethodInfo*);
 				    *(HtVector3f*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -7076,7 +7044,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef HtVector4f(*_NativeMethod_)(MethodInfo*);
 				    *(HtVector4f*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)(_resolvedMethod);
 				    ip += 8;
@@ -7087,7 +7055,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(int32_t, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 8;
@@ -7099,7 +7067,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(int32_t, int32_t, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7112,7 +7080,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(int32_t, int32_t, int32_t, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7126,7 +7094,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(int32_t, int32_t, int32_t, int32_t, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), (*(int32_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7137,7 +7105,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(int64_t, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 8;
@@ -7149,7 +7117,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(int64_t, int64_t, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7162,7 +7130,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(int64_t, int64_t, int64_t, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7176,7 +7144,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(int64_t, int64_t, int64_t, int64_t, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), (*(int64_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7187,7 +7155,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(float, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 8;
@@ -7199,7 +7167,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(float, float, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7212,7 +7180,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(float, float, float, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7226,7 +7194,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(float, float, float, float, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), (*(float*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7237,7 +7205,7 @@ else \
 					uint32_t __method = *(uint32_t*)(ip + 4);
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(double, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 8;
@@ -7249,7 +7217,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(double, double, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7262,7 +7230,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(double, double, double, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7276,7 +7244,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef void(*_NativeMethod_)(double, double, double, double, MethodInfo*);
 				    ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), (*(double*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7288,7 +7256,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(int32_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7301,7 +7269,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(int32_t, int32_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7315,7 +7283,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(int32_t, int32_t, int32_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7330,7 +7298,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(int32_t, int32_t, int32_t, int32_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), (*(int32_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7342,7 +7310,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(int64_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7355,7 +7323,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(int64_t, int64_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7369,7 +7337,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(int64_t, int64_t, int64_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7384,7 +7352,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(int64_t, int64_t, int64_t, int64_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), (*(int64_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7396,7 +7364,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(float, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7409,7 +7377,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(float, float, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7423,7 +7391,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(float, float, float, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7438,7 +7406,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(float, float, float, float, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), (*(float*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7450,7 +7418,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(double, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7463,7 +7431,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(double, double, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7477,7 +7445,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(double, double, double, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7492,7 +7460,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef uint8_t(*_NativeMethod_)(double, double, double, double, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), (*(double*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7504,7 +7472,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(int32_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7517,7 +7485,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(int32_t, int32_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7531,7 +7499,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(int32_t, int32_t, int32_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7546,7 +7514,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(int32_t, int32_t, int32_t, int32_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), (*(int32_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7558,7 +7526,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(int64_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7571,7 +7539,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(int64_t, int64_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7585,7 +7553,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(int64_t, int64_t, int64_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7600,7 +7568,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(int64_t, int64_t, int64_t, int64_t, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), (*(int64_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7612,7 +7580,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(float, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7625,7 +7593,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(float, float, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7639,7 +7607,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(float, float, float, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7654,7 +7622,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(float, float, float, float, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), (*(float*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7666,7 +7634,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(double, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7679,7 +7647,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(double, double, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7693,7 +7661,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(double, double, double, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7708,7 +7676,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int32_t(*_NativeMethod_)(double, double, double, double, MethodInfo*);
 				    *(int32_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), (*(double*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7720,7 +7688,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(int32_t, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7733,7 +7701,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(int32_t, int32_t, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7747,7 +7715,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(int32_t, int32_t, int32_t, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7762,7 +7730,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(int32_t, int32_t, int32_t, int32_t, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), (*(int32_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7774,7 +7742,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(int64_t, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7787,7 +7755,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(int64_t, int64_t, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7801,7 +7769,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(int64_t, int64_t, int64_t, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7816,7 +7784,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(int64_t, int64_t, int64_t, int64_t, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), (*(int64_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7828,7 +7796,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(float, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7841,7 +7809,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(float, float, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7855,7 +7823,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(float, float, float, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7870,7 +7838,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(float, float, float, float, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), (*(float*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7882,7 +7850,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(double, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7895,7 +7863,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(double, double, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7909,7 +7877,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(double, double, double, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7924,7 +7892,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef int64_t(*_NativeMethod_)(double, double, double, double, MethodInfo*);
 				    *(int64_t*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), (*(double*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7936,7 +7904,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(int32_t, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -7949,7 +7917,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(int32_t, int32_t, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -7963,7 +7931,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(int32_t, int32_t, int32_t, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -7978,7 +7946,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(int32_t, int32_t, int32_t, int32_t, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), (*(int32_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -7990,7 +7958,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(int64_t, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -8003,7 +7971,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(int64_t, int64_t, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -8017,7 +7985,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(int64_t, int64_t, int64_t, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -8032,7 +8000,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(int64_t, int64_t, int64_t, int64_t, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), (*(int64_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -8044,7 +8012,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(float, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -8057,7 +8025,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(float, float, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -8071,7 +8039,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(float, float, float, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -8086,7 +8054,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(float, float, float, float, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), (*(float*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -8098,7 +8066,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(double, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -8111,7 +8079,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(double, double, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -8125,7 +8093,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(double, double, double, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -8140,7 +8108,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef float(*_NativeMethod_)(double, double, double, double, MethodInfo*);
 				    *(float*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), (*(double*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -8152,7 +8120,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(int32_t, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -8165,7 +8133,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(int32_t, int32_t, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -8179,7 +8147,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(int32_t, int32_t, int32_t, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -8194,7 +8162,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(int32_t, int32_t, int32_t, int32_t, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int32_t*)(localVarBase + __param0)), (*(int32_t*)(localVarBase + __param1)), (*(int32_t*)(localVarBase + __param2)), (*(int32_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -8206,7 +8174,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(int64_t, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -8219,7 +8187,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(int64_t, int64_t, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -8233,7 +8201,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(int64_t, int64_t, int64_t, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -8248,7 +8216,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(int64_t, int64_t, int64_t, int64_t, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(int64_t*)(localVarBase + __param0)), (*(int64_t*)(localVarBase + __param1)), (*(int64_t*)(localVarBase + __param2)), (*(int64_t*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -8260,7 +8228,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(float, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -8273,7 +8241,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(float, float, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -8287,7 +8255,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(float, float, float, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -8302,7 +8270,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(float, float, float, float, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(float*)(localVarBase + __param0)), (*(float*)(localVarBase + __param1)), (*(float*)(localVarBase + __param2)), (*(float*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -8314,7 +8282,7 @@ else \
 					uint16_t __param0 = *(uint16_t*)(ip + 2);
 					uint16_t __ret = *(uint16_t*)(ip + 4);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(double, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), _resolvedMethod);
 				    ip += 16;
@@ -8327,7 +8295,7 @@ else \
 					uint16_t __param1 = *(uint16_t*)(ip + 4);
 					uint16_t __ret = *(uint16_t*)(ip + 6);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(double, double, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), _resolvedMethod);
 				    ip += 16;
@@ -8341,7 +8309,7 @@ else \
 					uint16_t __param2 = *(uint16_t*)(ip + 6);
 					uint16_t __ret = *(uint16_t*)(ip + 8);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(double, double, double, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), _resolvedMethod);
 				    ip += 16;
@@ -8356,7 +8324,7 @@ else \
 					uint16_t __param3 = *(uint16_t*)(ip + 8);
 					uint16_t __ret = *(uint16_t*)(ip + 10);
 				    MethodInfo* _resolvedMethod = ((MethodInfo*)imi->resolveDatas[__method]);
-				    Interpreter::RuntimeClassCCtorInit(_resolvedMethod);
+				    RuntimeClassCCtorInit(_resolvedMethod);
 				    typedef double(*_NativeMethod_)(double, double, double, double, MethodInfo*);
 				    *(double*)(void*)(localVarBase + __ret) = ((_NativeMethod_)_resolvedMethod->methodPointer)((*(double*)(localVarBase + __param0)), (*(double*)(localVarBase + __param1)), (*(double*)(localVarBase + __param2)), (*(double*)(localVarBase + __param3)), _resolvedMethod);
 				    ip += 16;
@@ -9329,7 +9297,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(int8_t*)(((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9339,7 +9307,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(uint8_t*)(((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9349,7 +9317,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(int16_t*)(((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9359,7 +9327,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(uint16_t*)(((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9369,7 +9337,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(int32_t*)(((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9379,7 +9347,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(uint32_t*)(((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9389,7 +9357,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int64_t*)(localVarBase + __dst)) = *(int64_t*)(((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9399,7 +9367,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int64_t*)(localVarBase + __dst)) = *(uint64_t*)(((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9409,7 +9377,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy8((void*)(localVarBase + __dst), ((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9419,7 +9387,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy12((void*)(localVarBase + __dst), ((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9429,7 +9397,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy16((void*)(localVarBase + __dst), ((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9439,7 +9407,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy20((void*)(localVarBase + __dst), ((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9449,7 +9417,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy24((void*)(localVarBase + __dst), ((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9459,7 +9427,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy28((void*)(localVarBase + __dst), ((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9469,7 +9437,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy32((void*)(localVarBase + __dst), ((byte*)__klass->static_fields) + __offset);
 				    ip += 16;
 				    continue;
@@ -9480,7 +9448,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
 					uint16_t __size = *(uint16_t*)(ip + 6);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    std::memmove((void*)(localVarBase + __dst), (((byte*)__klass->static_fields) + __offset), __size);
 				    ip += 16;
 				    continue;
@@ -9491,7 +9459,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
 					uint32_t __size = *(uint32_t*)(ip + 12);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    std::memmove((void*)(localVarBase + __dst), (((byte*)__klass->static_fields) + __offset), __size);
 				    ip += 16;
 				    continue;
@@ -9501,7 +9469,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(int8_t*)(((byte*)__klass->static_fields) + __offset) = (*(int8_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9511,7 +9479,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(uint8_t*)(((byte*)__klass->static_fields) + __offset) = (*(uint8_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9521,7 +9489,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(int16_t*)(((byte*)__klass->static_fields) + __offset) = (*(int16_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9531,7 +9499,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(uint16_t*)(((byte*)__klass->static_fields) + __offset) = (*(uint16_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9541,7 +9509,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(int32_t*)(((byte*)__klass->static_fields) + __offset) = (*(int32_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9551,7 +9519,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(uint32_t*)(((byte*)__klass->static_fields) + __offset) = (*(uint32_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9561,7 +9529,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(int64_t*)(((byte*)__klass->static_fields) + __offset) = (*(int64_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9571,7 +9539,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(uint64_t*)(((byte*)__klass->static_fields) + __offset) = (*(uint64_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9581,7 +9549,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy8(((byte*)__klass->static_fields) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9591,7 +9559,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy12(((byte*)__klass->static_fields) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9601,7 +9569,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy16(((byte*)__klass->static_fields) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9611,7 +9579,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy20(((byte*)__klass->static_fields) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9621,7 +9589,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy24(((byte*)__klass->static_fields) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9631,7 +9599,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy28(((byte*)__klass->static_fields) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9641,7 +9609,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy32(((byte*)__klass->static_fields) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9652,7 +9620,7 @@ else \
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
 					uint16_t __size = *(uint16_t*)(ip + 6);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    std::memmove(((byte*)__klass->static_fields) + __offset, (void*)(localVarBase + __data), __size);
 				    ip += 16;
 				    continue;
@@ -9663,7 +9631,7 @@ else \
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
 					uint32_t __size = *(uint32_t*)(ip + 12);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    std::memmove(((byte*)__klass->static_fields) + __offset, (void*)(localVarBase + __data), __size);
 				    ip += 16;
 				    continue;
@@ -9673,7 +9641,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(void**)(localVarBase + __dst)) = ((byte*)__klass->static_fields) + __offset;
 				    ip += 16;
 				    continue;
@@ -9691,7 +9659,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(void**)(localVarBase + __dst)) = (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset;
 				    ip += 16;
 				    continue;
@@ -9701,7 +9669,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(int8_t*)((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9711,7 +9679,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(uint8_t*)((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9721,7 +9689,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(int16_t*)((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9731,7 +9699,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(uint16_t*)((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9741,7 +9709,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(int32_t*)((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9751,7 +9719,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int32_t*)(localVarBase + __dst)) = *(uint32_t*)((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9761,7 +9729,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int64_t*)(localVarBase + __dst)) = *(int64_t*)((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9771,7 +9739,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    (*(int64_t*)(localVarBase + __dst)) = *(uint64_t*)((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9781,7 +9749,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy8((void*)(localVarBase + __dst), (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9791,7 +9759,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy12((void*)(localVarBase + __dst), (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9801,7 +9769,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy16((void*)(localVarBase + __dst), (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9811,7 +9779,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy20((void*)(localVarBase + __dst), (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9821,7 +9789,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy24((void*)(localVarBase + __dst), (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9831,7 +9799,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy28((void*)(localVarBase + __dst), (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9841,7 +9809,7 @@ else \
 					uint16_t __dst = *(uint16_t*)(ip + 2);
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy32((void*)(localVarBase + __dst), (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset);
 				    ip += 16;
 				    continue;
@@ -9852,7 +9820,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					int32_t __offset = *(int32_t*)(ip + 12);
 					uint16_t __size = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    std::memmove((void*)(localVarBase + __dst), (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset, __size);
 				    ip += 16;
 				    continue;
@@ -9863,7 +9831,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 4)]);
 					int32_t __offset = *(int32_t*)(ip + 8);
 					uint32_t __size = *(uint32_t*)(ip + 12);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    std::memmove((void*)(localVarBase + __dst), (byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset, __size);
 				    ip += 16;
 				    continue;
@@ -9873,7 +9841,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(int8_t*)((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset) = (*(int8_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9883,7 +9851,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(uint8_t*)((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset) = (*(uint8_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9893,7 +9861,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(int16_t*)((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset) = (*(int16_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9903,7 +9871,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(uint16_t*)((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset) = (*(uint16_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9913,7 +9881,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(int32_t*)((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset) = (*(int32_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9923,7 +9891,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(uint32_t*)((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset) = (*(uint32_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9933,7 +9901,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(int64_t*)((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset) = (*(int64_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9943,7 +9911,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    *(uint64_t*)((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset) = (*(uint64_t*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9953,7 +9921,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy8((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9963,7 +9931,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy12((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9973,7 +9941,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy16((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9983,7 +9951,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy20((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -9993,7 +9961,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy24((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -10003,7 +9971,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy28((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -10013,7 +9981,7 @@ else \
 					Il2CppClass* __klass = ((Il2CppClass*)imi->resolveDatas[*(uint32_t*)(ip + 8)]);
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    Copy32((byte*)(il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset)) + __offset, (void*)(localVarBase + __data));
 				    ip += 16;
 				    continue;
@@ -10024,7 +9992,7 @@ else \
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
 					uint16_t __size = *(uint16_t*)(ip + 6);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    std::memmove((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset, (void*)(localVarBase + __data), __size);
 				    ip += 16;
 				    continue;
@@ -10035,7 +10003,7 @@ else \
 					uint16_t __offset = *(uint16_t*)(ip + 2);
 					uint16_t __data = *(uint16_t*)(ip + 4);
 					uint32_t __size = *(uint32_t*)(ip + 12);
-				    Interpreter::RuntimeClassCCtorInit(__klass);
+				    RuntimeClassCCtorInit(__klass);
 				    std::memmove((byte*)il2cpp::vm::Thread::GetThreadStaticData(__klass->thread_static_fields_offset) + __offset, (void*)(localVarBase + __data), __size);
 				    ip += 16;
 				    continue;
