+++ hybridclr/interpreter/MethodBridge.h
@@ -67,73 +67,9 @@ namespace interpreter
 		uint64_t high;
 	};
 
-	inline bool IsNeedExpandLocationType(LocationDataType type)
-	{
-		return type < LocationDataType::U8;
-	}
-
-	ArgDesc GetTypeArgDesc(const Il2CppType* type);
-
-	inline LocationDataType GetLocationDataTypeByType(const Il2CppType* type)
-	{
-		return GetTypeArgDesc(type).type;
-	}
-
-	inline void ExpandLocationData2StackDataByType(void* retValue, LocationDataType type)
-	{
-		switch (type)
-		{
-		case hybridclr::interpreter::LocationDataType::I1:
-			*(int32_t*)retValue = *(int8_t*)retValue;
-			break;
-		case hybridclr::interpreter::LocationDataType::U1:
-			*(int32_t*)retValue = *(uint8_t*)retValue;
-			break;
-		case hybridclr::interpreter::LocationDataType::I2:
-			*(int32_t*)retValue = *(int16_t*)retValue;
-			break;
-		case hybridclr::interpreter::LocationDataType::U2:
-			*(int32_t*)retValue = *(uint16_t*)retValue;
-			break;
-		default:
-			break;
-		}
-	}
-
-	inline void ExpandLocationData2StackDataByType(void* retValue, Il2CppTypeEnum type)
-	{
-		switch (type)
-		{
-		case IL2CPP_TYPE_BOOLEAN:
-		case IL2CPP_TYPE_I1:
-			*(int32_t*)retValue = *(int8_t*)retValue;
-			break;
-		case IL2CPP_TYPE_U1:
-			*(int32_t*)retValue = *(uint8_t*)retValue;
-			break;
-		case IL2CPP_TYPE_I2:
-			*(int32_t*)retValue = *(int16_t*)retValue;
-			break;
-		case IL2CPP_TYPE_U2:
-		case IL2CPP_TYPE_CHAR:
-			*(int32_t*)retValue = *(uint16_t*)retValue;
-			break;
-		default:
-			break;
-		}
-	}
-
-	ArgDesc GetValueTypeArgDescBySize(uint32_t size);
-
-	inline bool IsSimpleStackObjectCopyArg(LocationDataType type)
-	{
-		return type <= LocationDataType::U8;
-	}
-
 	void CopyArgs(StackObject* dstBase, StackObject* argBase, ArgDesc* args, uint32_t paramCount, uint32_t totalParamStackObjectSize);
 
 	bool IsPassArgAsValue(const Il2CppType* type, LocationDataType* locType = nullptr);
-	Il2CppObject* TranslateNativeValueToBoxValue(const Il2CppType* type, void* value);
 	void ConvertInvokeArgs(StackObject* resultArgs, const MethodInfo* method, void** __args);
 
 	bool ComputeSignature(const MethodInfo* method, bool call, char* sigBuf, size_t bufferSize);
